moved reference-store to website-bundle
