+++ src/Sulu/Bundle/HttpCacheBundle/Resources/config/structure-cache-handlers.xml
@@ -12,7 +12,7 @@
 
         <service id="sulu_http_cache.handler.tags" class="Sulu\Component\HttpCache\Handler\TagsHandler">
             <argument type="service" id="sulu_http_cache.proxy_client"/>
-            <argument type="service" id="sulu_content.reference_store_pool"/>
+            <argument type="service" id="sulu_website.reference_store_pool"/>
             <tag name="sulu_http_cache.handler" alias="tags" />
         </service>
