+++ src/Sulu/Bundle/HttpCacheBundle/Tests/Unit/DependencyInjection/SuluHttpCacheExtensionTest.php
@@ -13,8 +13,8 @@ namespace Sulu\Bundle\HttpCacheBundle\Tests\Unit\DependencyInjection;
 
 use Matthias\SymfonyDependencyInjectionTest\PhpUnit\AbstractExtensionTestCase;
 use Psr\Log\LoggerInterface;
-use Sulu\Bundle\ContentBundle\ReferenceStore\ReferenceStorePoolInterface;
 use Sulu\Bundle\HttpCacheBundle\DependencyInjection\SuluHttpCacheExtension;
+use Sulu\Bundle\WebsiteBundle\ReferenceStore\ReferenceStorePoolInterface;
 use Sulu\Component\Content\ContentTypeManagerInterface;
 use Sulu\Component\Webspace\Manager\WebspaceManagerInterface;
 use Sulu\Component\Webspace\Url\ReplacerInterface;
@@ -69,7 +69,7 @@ class SuluHttpCacheExtensionTest extends AbstractExtensionTestCase
         $this->container->set('request_stack', $this->requestStack->reveal());
         $this->container->set('sulu_core.webspace.webspace_manager.url_replacer', $this->replacer->reveal());
         $this->container->set('logger', $this->logger->reveal());
-        $this->container->set('sulu_content.reference_store_pool', $this->referenceStore->reveal());
+        $this->container->set('sulu_website.reference_store_pool', $this->referenceStore->reveal());
     }
 
     protected function getContainerExtensions()
