+++ src/Sulu/Bundle/SnippetBundle/Resources/config/content.xml
@@ -5,8 +5,8 @@
 
     <services>
         <service id="sulu_snippet.reference_store.snippet"
-                 class="Sulu\Bundle\ContentBundle\ReferenceStore\ReferenceStore">
-            <tag name="sulu_content.reference_store" alias="snippet"/>
+                 class="Sulu\Bundle\WebsiteBundle\ReferenceStore\ReferenceStore">
+            <tag name="sulu_website.reference_store" alias="snippet"/>
         </service>
 
         <service id="sulu_snippet.default_snippet.manager" class="Sulu\Bundle\SnippetBundle\Snippet\DefaultSnippetManager">
